# src/main/resources/openapi/contribution-api-spec.yaml
---
openapi: 3.0.3
info:
  version: 1.0.0
  title: Contributions API
  description: OpenAPI specification for Contribution APIs
  termsOfService: https://changeme.optum.com/
  contact:
    name: Optum Financial OBI Team
    email: Team_OBI@ds.uhc.com
    url: https://www.optum.com/
paths:
  /contributionSummary:
    post:
      operationId: getContributionSummary
      summary: Get Contribution Summary
      tags:
      - Contribution
      - Summary
      requestBody:
        $ref: "#/components/requestBodies/ContributionSummaryRequest"
      responses:
        "200":
          $ref: "#/components/responses/ContributionSummaryResponse"
        "400":
          $ref: "#/components/responses/BadRequestErrorResponse"
        "404":
          $ref: "#/components/responses/NotFoundErrorResponse"
        "500":
          $ref: "#/components/responses/ServerErrorResponse"
        default:
          $ref: "#/components/responses/UnexpectedErrorResponse"
  /contributionDetails:
    post:
      operationId: getContributionDetails
      summary: Get Contribution Details
      tags:
      - Contribution
      - Details
      requestBody:
        $ref: "#/components/requestBodies/ContributionDetailsRequest"
      responses:
        "200":
          $ref: "#/components/responses/ContributionDetailsResponse"
        "400":
          $ref: "#/components/responses/BadRequestErrorResponse"
        "404":
          $ref: "#/components/responses/NotFoundErrorResponse"
        "500":
          $ref: "#/components/responses/ServerErrorResponse"
        default:
          $ref: "#/components/responses/UnexpectedErrorResponse"
  /updateContributionFileStatus:
    patch:
      operationId: updateContributionFileStatus
      summary: Update the status of a contribution file.
      tags:
      - Contributions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContributionFileStatusRequest"
      responses:
        "200":
          description: Status response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateContributionFileStatusResponse"
  /getEmployerContributionFileHistory:
    get:
      operationId: getEmployerContributionFileHistory
      summary: Retrieve the history of employer contribution files.
      tags:
      - Contributions
      parameters:
      - $ref: "#/components/parameters/QueryCustomerPolicyNumber"
      - $ref: "#/components/parameters/QueryEffectiveTransactionDate"
      responses:
        "200":
          description: Contribution file history response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionFileHistoryResponse"
  /getEmployerContributionFileTransactions:
    get:
      operationId: getEmployerContributionFileTransactions
      summary: Retrieve transactions for employer contribution files.
      tags:
      - Contributions
      parameters:
      - $ref: "#/components/parameters/QueryContributionFilePortalId"
      responses:
        "200":
          description: Contribution file transactions response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionFileTransactionResponse"
  /getEmployerContributionFileNames:
    get:
      operationId: getEmployerContributionFileNames
      summary: Retrieve names of employer contribution files.
      tags:
      - Contributions
      parameters:
      - $ref: "#/components/parameters/QueryContributionFilePortalIds"
      responses:
        "200":
          description: Contribution file names response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionFileNamesResponse"
components:
  requestBodies:
    ContributionSummaryRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContributionSummaryRequest"
    ContributionDetailsRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContributionDetailsRequest"
  responses:
    BadRequestErrorResponse:
      description: Invalid element on request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundErrorResponse:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerErrorResponse:
      description: Server failed to process the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnexpectedErrorResponse:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ContributionSummaryResponse:
      description: valid response upon successful summary retrieval
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContributionSummaryResponse"
    ContributionDetailsResponse:
      description: valid response upon successful details retrieval
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContributionDetailsResponse"
  parameters:
    QueryCustomerPolicyNumber:
      in: query
      name: customerPolicyNumber
      required: true
      schema:
        type: string
        description: Customer policy number
    QueryEffectiveTransactionDate:
      in: query
      name: effectiveTransactionDate
      required: true
      schema:
        type: string
        description: Effective transaction date
    QueryContributionFilePortalId:
      in: query
      name: contributionFilePortalId
      required: true
      schema:
        type: string
        description: Portal Id for contribution file
    QueryContributionFilePortalIds:
      in: query
      name: contributionFilePortalId
      required: true
      explode: true
      schema:
        type: array
        items:
          type: string
        description: List of portal Ids for contribution file
  schemas:
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
    ErrorItem:
      type: object
      properties:
        statusCode:
          type: string
          description: The status code based on the type of error.
          example: 2
        severity:
          type: string
          example: ERR
        statusDescription:
          type: string
          example: INCORRECT INPUT FORMAT
    ContributionSummaryRequest:
      type: object
      required:
      - employerGroupId
      - restrictToFunded
      - dateFrom
      - dateTo
      properties:
        employerGroupId:
          type: string
        restrictToFunded:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        maxNumberOfFiles:
          type: integer
        requestId:
          type: string
    ContributionSummaryResponse:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/Status"
        requestId:
          type: string
        contributionFileSummary:
          type: array
          items:
            $ref: "#/components/schemas/ContributionFileSummary"
    ContributionFileSummary:
      type: object
      properties:
        fileStatus:
          $ref: "#/components/schemas/Status"
        contributionFileInfo:
          $ref: "#/components/schemas/ContributionFileInfo"
    ContributionFileInfo:
      type: object
      properties:
        contributionFileKey:
          type: string
        employerGroupNumber:
          type: string
        fundedIndicator:
          type: string
        submittedDate:
          type: string
        effectiveDate:
          type: string
        validRecordCount:
          type: number
        totalValidContributionAmount:
          type: string
        invalidRecordCount:
          type: number
        invalidTotalValidContributionAmount:
          type: number
    Status:
      type: object
      properties:
        statusCode:
          type: string
        severity:
          type: string
        statusDescription:
          type: string
    ContributionDetailsRequest:
      type: object
      required:
      - fileDetails
      properties:
        requestId:
          type: string
        fileDetails:
          type: array
          items:
            $ref: "#/components/schemas/FileDetails"
    FileDetails:
      type: object
      properties:
        contributionFileId:
          type: string
        requestType:
          type: string
    ContributionDetailsResponse:
      type: object
      properties:
        requestId:
          type: string
        contrbDetailStatus:
          $ref: "#/components/schemas/ContributionDetailsStatus"
        contributionFiles:
          type: array
          items:
            $ref: "#/components/schemas/ContributionFile"
    ContributionFile:
      type: object
      properties:
        contributionFileId:
          type: string
        contributionFileStatus:
          $ref: "#/components/schemas/ContributionDetailsStatus"
        contributionRecords:
          type: array
          items:
            $ref: "#/components/schemas/ContributionRecord"
    ContributionRecord:
      type: object
      properties:
        accountHolderSsn:
          type: string
        personName:
          $ref: "#/components/schemas/PersonName"
        contributionAmount:
          type: number
        contributionType:
          $ref: "#/components/schemas/ContributionType"
        validIndicator:
          type: string
        invalidReasons:
          type: array
          items:
            $ref: "#/components/schemas/InvalidReason"
    PersonName:
      type: object
      properties:
        personFirstName:
          type: string
        personMiddleName:
          type: string
        personLastName:
          type: string
    ContributionType:
      type: object
      properties:
        code:
          type: string
        codeName:
          type: string
    InvalidReason:
      type: object
      properties:
        code:
          type: string
        codeName:
          type: string
    ContributionDetailsStatus:
      type: object
      properties:
        statusCode:
          type: string
        severity:
          type: string
        statusDescription:
          type: string
    UpdateContributionFileStatusRequest:
      type: object
      properties:
        contributionFileId:
          type: string
        customerFundId:
          type: string
        effectiveTransactionDate:
          type: string
        changedByName:
          type: string
        commentText:
          type: string
        fileStatusCode:
          type: string
    ContributionFileHistoryResponse:
      type: object
      properties:
        contributionFileHistoryRecords:
          type: array
          items:
            $ref: "#/components/schemas/ContributionFileHistoryRecords"
    ContributionFileHistoryRecords:
      type: object
      properties:
        approvedByName:
          type: string
        approvedTimestamp:
          type: string
        approverEmailAndTimestamp:
          type: string
        bankNickName:
          type: string
        bankName:
          type: string
        contributionFilePortalId:
          type: string
        customerFundId:
          type: string
        customerPolicyNumber:
          type: string
        denyByName:
          type: string
        denyCommitText:
          type: string
        denyTimestamp:
          type: string
        effectiveTransactionDate:
          type: string
        friendlyFileName:
          type: string
        newFileName:
          type: string
        originalFileName:
          type: string
        processedTimestamp:
          type: string
        submittedByName:
          type: string
        submittedMethodCode:
          type: string
        submittedTimestamp:
          type: string
        statusCode:
          type: string
        totalFileContributionAmount:
          type: string
    ContributionFileTransactionResponse:
      type: object
      properties:
        contributionFiles:
          type: array
          items:
            $ref: "#/components/schemas/ContributionFileItem"
    ContributionFileItem:
      type: object
      properties:
        accountNumber:
          type: string
        contributionFilePortalId:
          type: string
        contributionTypeName:
          type: string
        crAmt:
          type: string
        portalContributionCommitText:
          type: string
        ssn:
          type: string
        transactionDescription:
          type: string
    ContributionFileNamesResponse:
      type: object
      properties:
        contributionFileNames:
          type: array
          items:
            $ref: "#/components/schemas/ContributionFileName"
    ContributionFileName:
      type: object
      properties:
        contributionFilePortalId:
          type: string
        effectiveTransactionDate:
          type: string
        friendlyFileName:
          type: string
        newFileName:
          type: string
        originalFileName:
          type: string
        processedTimestamp:
          type: string
    UpdateContributionFileStatusResponse:
      type: object
      properties:
        contributionFileId:
          type: string